refactor : repackage app
