+++ src/main/java/app/nextstep/dao/MemberDao.java
@@ -1,5 +1,6 @@
-package app.nextstep.member;
+package app.nextstep.dao;
 
+import app.nextstep.domain.Member;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
