+++ src/test/java/app/auth/AuthE2ETest.java
@@ -4,8 +4,8 @@ import app.auth.dto.TokenRequest;
 import app.auth.dto.TokenResponse;
 import app.nextstep.RoomEscapeApplication;
 import io.restassured.RestAssured;
-import app.nextstep.member.MemberRequest;
-import app.nextstep.theme.ThemeRequest;
+import app.nextstep.dto.MemberRequest;
+import app.nextstep.dto.ThemeRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
