+++ src/test/java/app/nextstep/theme/ThemeDaoTest.java
@@ -1,5 +1,7 @@
 package app.nextstep.theme;
 
+import app.nextstep.dao.ThemeDao;
+import app.nextstep.domain.Theme;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
